requirements.txt
# Web Framework & API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.1

# Monitoring & Logging
loguru>=0.7.2
prometheus-fastapi-instrumentator>=6.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1   # passlib 1.7.4 is incompatible with bcrypt>=5.0

# Frontend
streamlit>=1.33.0
plotly>=5.20.0

# HTTP Client
requests>=2.31.0

# ML / Data
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.4.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.27.0